backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
alembic
psycopg2-binary
asyncpg
python-dotenv
twilio
pydantic
pydantic-settings
python-multipart
email-validator
aiosqlite
deepgram-sdk<4
websockets
redis
openai
faiss-cpu
numpy
pinecone-client
pgvector
psycopg2
tiktoken
pypdf2
python-docx
langchain
langchain-openai
langchain-community
groq
sentence-transformers
passlib[bcrypt]
python-jose[cryptography]
